+++ R/hr_documentation.R
@@ -6,9 +6,10 @@
 #' area of an home range, `hr_isopleths` returns the isopleth as a
 #' `SpatialPolygonsDataFrame`.
 #'
-#' The implementation of the reference bandwidth calculation is based on Worton (1989). If variances differ greatly, it is advaisable to rescale the data using `rescale = "unitvar"` the data is suspected to multimodal other bandwidth estimation methods may be more suitable.
+#' The implementation of the reference bandwidth calculation is based on Worton (1989). If variances differ greatly, it is advisable to rescale the data using `rescale = "unitvar"` the data is suspected to multimodal other bandwidth estimation methods may be more suitable.
 #'
 #' @template track_xy_star
+#' @param units `[logic(1)]` \cr Should areas be returned as units? If `FALSE` areas are returned as numeric values.
 #' @param levels `[numeric]` \cr The isopleth levels used for calculating home
 #'   ranges. Should be `0 < level < 1`.
 #' @param n `[integer(1)]` \cr The number of neighbors used when calculating
@@ -17,7 +18,8 @@
 #'   area 0 (if coordinates are numerically identical).
 #' @param trast `[RasterLayer]` \cr A template raster for kernel density home-ranges.
 #' @param h `[numeric(2)]` \cr The bandwidth for kernel density estimation.
-#' @param rescale `[character(1)]` \cr Rescaling method for reference bandwidth calucation. Must be one of "unitvar", "xvar", or "none".
+#' @param keep.data `[logic(2)]` \cr Should the original tracking data be included in the estimate?
+#' @param rescale `[character(1)]` \cr Rescaling method for reference bandwidth calculation. Must be one of "unitvar", "xvar", or "none".
 #' @template dots_none
 #' @references Worton, B. J. (1989). Kernel methods for estimating the utilization distribution in home-range studies. _Ecology, 70(1)_, 164-168.
 #'