+++ R/hr_locoh.R
@@ -8,7 +8,7 @@ hr_locoh <- function(x, ...) {
 #' @export
 #' @param type `k`, `r` or `a`. Type of LoCoH.
 #' @rdname hr
-hr_locoh.track_xy <- function(x, n = 10, type = "k", levels = 0.95, rand_buffer = 1e-5, ...) {
+hr_locoh.track_xy <- function(x, n = 10, type = "k", levels = 0.95, keep.data = TRUE, rand_buffer = 1e-5, ...) {
 
   ## type
   if (!type %in% c("a", "k", "r")) {
@@ -30,6 +30,8 @@ hr_locoh.track_xy <- function(x, n = 10, type = "k", levels = 0.95, rand_buffer 
     ## 2. order by dist
     ## 3. take n nearest
     aa <- FNN::get.knn(x[, c("x_", "y_")], k = n)$nn.index
+    #"r" and "a" methods return a list -- split so that this is the same
+    aa <- split(aa, 1:nrow(aa))
   } else if (type == "r") {
     ## 1. calc dist
     ## 2. take all pts with dist <= n
@@ -41,13 +43,13 @@ hr_locoh.track_xy <- function(x, n = 10, type = "k", levels = 0.95, rand_buffer 
     # 3. take cum dist
     # 4. take points where cumist <= n
     aa <- lapply(no, function(i) {
-      di <- sqrt((x$x_ - x$y_[i])^2 + (x$y_ - x$y_[i])^2)
+      di <- sqrt((x$x_ - x$x_[i])^2 + (x$y_ - x$y_[i])^2)
       no[order(di)][cumsum(di[order(di)]) <= n]
     })
   }
 
   xysp <- sp::SpatialPointsDataFrame(x[, c("x_", "y_")], data=data.frame(id=1:nrow(x)))
-  zz <- lapply(1:nrow(aa), function(i) xysp[aa[i, ], ])
+  zz <- lapply(1:length(aa), function(i) xysp[aa[[i]], ])
   mcps <- lapply(zz, function(x) rgeos::gBuffer(rgeos::gConvexHull(x), width = rand_buffer))
   mcpAreas <- sapply(mcps, rgeos::gArea)
 
@@ -98,7 +100,12 @@ hr_locoh.track_xy <- function(x, n = 10, type = "k", levels = 0.95, rand_buffer 
     sp::proj4string(qq2) <- attr(x, "crs_")
   }
 
-  out <- list(locoh = qq2)
+  qq2 <- sf::st_as_sf(qq2)
+  qq2$area <- sf::st_area(qq2)
+
+  out <- list(locoh = qq2, levels = levels, type = type, n = n, estimator = "locoh",
+              crs = get_crs(x),
+              data = if (keep.data) x else NULL)
   class(out) <- c("locoh", "hr_geom", "hr", "list")
   out