+++ NEWS.md
@@ -1,7 +1,37 @@
+# amt 0.1.3
+## bug fixes
+
+## updates
+- `hr_overlap` gained two new arguemnts `labels` and `consecutive.only`.
+
+## bug fixes
+- `group_by` for random points works as expected.
+
+# amt 0.1.2
+## bug fixes
+
+## updates
+- `hr_mcp` gained an argument `keep.data`
+- `as_track` can now convert `steps_xyt` back to a `track_xyt`
+- `plot` function for home ranges improved
+- Ported function for `site_fidelity` from package `rhr`.
+- Changed test in order to be compatible with the new dplyr version. 
+
+
+# amt 0.1.0
+## bug fixes
+- `random_steps` uses abs direction as reference instead of relative direction
+
+## updates
+- Streamlined home-range methods
+- Added methods for overlaps and intersections of home ranges
+- Added href scaled for KDE
+
+
 # amt 0.0.9
 ## bug fixes
 - `extract_covariate` now works with new version of tibble for raster stacks.
-- `fit_distr` for gamma distribution, now replaces 0 step length with min setep- length and informs the user with a message. 
+- `fit_distr` for gamma distribution, now replaces 0 step length with min step- length and informs the user with a message. 
 
 ## minor changes
 - added citation to the description
@@ -20,7 +50,7 @@
 
 
 ## major changes
-- Improved `random_points` by addeing methods for track, hr and spatial objects.
+- Improved `random_points` by adding methods for track, hr and spatial objects.
 - Adjusted `random_steps`
 - Removed dependency on `velox`. This makes `extract_covariates_along` slower. 
 
@@ -55,14 +85,14 @@
 - bug fix in `random_steps`
 - new function `extract_covariates_along`, extracts covariates along a random step.
 - `steps` gains a new argument: `keep_cols` in order to keep columns from point when creating steps.
-- `simulate_tud`: a convenience wrapper arround `simulate_ud` to simulate transition UDs.
+- `simulate_tud`: a convenience wrapper around `simulate_ud` to simulate transition UDs.
 # amt 0.0.4.0
 ## major changes
 - Added an alias to `mk_track` called `make_track`.
 - `mk_track` gained a new argument `all_cols` that allows to carry over all columns to track.
 - `mk_track` gained a new argument `check_duplicates` that allows to the user to specify if temporal duplicates should give an error or not.
 - Added `hr_kde` to calculate Kernel home ranges.
-- `steps` now uses `difftime` to calcualte time differences.
+- `steps` now uses `difftime` to calculate time differences.
 
 ## minor changes
 - Removed: Depends tidyverse
@@ -71,5 +101,5 @@
 # amt 0.0.3.0
 ## minor changes
 
-- Fixed vignette titels.
+- Fixed vignette titles.
 - Added `Rdpack` to suggest